requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
requests
python-dotenv
fuzzywuzzy
python-Levenshtein
sentence-transformers
sqlalchemy
psycopg2-binary
mysql-connector-python